# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SkillQuest",
    "description": "Игровой бэкенд SkillQuest: диагностика навыков, игровые миссии, рекомендации курсов и стажировок, анализ резюме. Защищённые операции требуют Bearer-токен (см. security и docs/API.md).",
    "contact": {
      "name": "SkillQuest Team",
      "url": "https://github.com/BrainL4g/max-edu-2026"
    },
    "version": "0.2.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
    "/users": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Create User",
        "description": "Создание пользователя (доступно любому аутентифицированному).",
        "operationId": "create_user_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "401": {
            "description": "Нет или невалидный Bearer-токен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/users/by-max": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Get Or Create By Max",
        "description": "Связка пользователя MAX: get-or-create по max_user_id (только service).",
        "operationId": "get_or_create_by_max_users_by_max_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserByMaxIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "401": {
            "description": "Нет или невалидный Bearer-токен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён: не свой ресурс или не та роль",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/users/{user_id}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get User",
        "description": "Профиль пользователя.",
        "operationId": "get_user_users__user_id__get",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "401": {
            "description": "Нет или невалидный Bearer-токен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён: не свой ресурс или не та роль",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "users"
        ],
        "summary": "Update User",
        "description": "Обновление профиля пользователя.",
        "operationId": "update_user_users__user_id__patch",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "401": {
            "description": "Нет или невалидный Bearer-токен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён: не свой ресурс или не та роль",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{user_id}/goal": {
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Set Goal",
        "description": "Выбор целевой роли пользователя.",
        "operationId": "set_goal_users__user_id__goal_put",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoalIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "401": {
            "description": "Нет или невалидный Bearer-токен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён: не свой ресурс или не та роль",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{user_id}/gap-analysis": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Gap Analysis",
        "description": "Что не хватает до целевой роли и процент соответствия.",
        "operationId": "gap_analysis_users__user_id__gap_analysis_get",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GapAnalysisOut"
                }
              }
            }
          },
          "401": {
            "description": "Нет или невалидный Bearer-токен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён: не свой ресурс или не та роль",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{user_id}/assessment": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Run Assessment",
        "description": "Первичная диагностика: ответы → начальный Skill Map.",
        "operationId": "run_assessment_users__user_id__assessment_post",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentOut"
                }
              }
            }
          },
          "401": {
            "description": "Нет или невалидный Bearer-токен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён: не свой ресурс или не та роль",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{user_id}/progress": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get Progress",
        "description": "Прогресс пользователя: XP, миссии, средний уровень, Skill Map.",
        "operationId": "get_progress_users__user_id__progress_get",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProgressOut"
                }
              }
            }
          },
          "401": {
            "description": "Нет или невалидный Bearer-токен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён: не свой ресурс или не та роль",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssessmentIn": {
        "properties": {
          "answers": {
            "items": {
              "$ref": "#/components/schemas/AssessmentAnswerIn"
            },
            "type": "array",
            "title": "Answers"
          }
        },
        "type": "object",
        "required": [
          "answers"
        ],
        "title": "AssessmentIn",
        "description": "Ответы на первичную диагностику.",
        "examples": [
          {
            "answers": [
              {
                "option_index": 2,
                "question_id": 1
              },
              {
                "option_index": 1,
                "question_id": 2
              }
            ]
          }
        ]
      },
      "AssessmentOut": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "evaluated_skills": {
            "items": {
              "$ref": "#/components/schemas/SkillMapItem"
            },
            "type": "array",
            "title": "Evaluated Skills"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "evaluated_skills",
          "summary"
        ],
        "title": "AssessmentOut",
        "description": "Результат диагностики: начальный Skill Map.",
        "examples": [
          {
            "evaluated_skills": [
              {
                "category": "Программирование",
                "experience": 0,
                "level": 2,
                "name": "Python",
                "next_level_xp": 50,
                "progress": 0.0,
                "skill_id": 1
              }
            ],
            "summary": "Ваш стартовый уровень — базовый.",
            "user_id": 1
          }
        ]
      },
...
      "ErrorOut": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorOut",
        "description": "Ошибка API: человекочитаемое сообщение в поле ``detail``."
      },
      "GapAnalysisOut": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RoleOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "match_percent": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Percent"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/GapItemOut"
            },
            "type": "array",
            "title": "Items"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "items",
          "summary"
        ],
        "title": "GapAnalysisOut",
        "description": "Gap-анализ пользователя: соответствие роли и пробелы.",
        "examples": [
          {
            "items": [
              {
                "category": "Программирование",
                "current_level": 2,
                "gap": 1,
                "importance": 1.0,
                "is_mandatory": true,
                "name": "Python",
                "required_level": 3,
                "skill_id": 1
              }
            ],
            "match_percent": 67,
            "role": {
              "direction": "backend",
              "id": 1,
              "level": "junior",
              "name": "Backend Junior",
              "requirements": []
            },
            "summary": "До целевой роли не хватает навыков: Python.",
            "user_id": 1
          }
        ]
      },
...
      "GoalIn": {
        "properties": {
          "target_role_id": {
            "type": "integer",
            "title": "Target Role Id"
          }
        },
        "type": "object",
        "required": [
          "target_role_id"
        ],
        "title": "GoalIn",
        "description": "Выбор целевой роли пользователем.",
        "examples": [
          {
            "target_role_id": 1
          }
        ]
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserByMaxIn": {
        "properties": {
          "max_user_id": {
            "type": "integer",
            "title": "Max User Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "max_user_id"
        ],
        "title": "UserByMaxIn",
        "description": "Связка пользователя MAX с платформой.",
        "examples": [
          {
            "max_user_id": 12345,
            "name": "Иван из MAX"
          }
        ]
      },
      "UserCreate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "education": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Education"
          },
          "direction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direction"
          },
          "goal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goal"
          }
        },
        "type": "object",
        "title": "UserCreate",
        "description": "Создание пользователя.",
        "examples": [
          {
            "direction": "backend",
            "education": "МГТУ, 3 курс",
            "goal": "Попасть на стажировку",
            "name": "Иван Иванов"
          }
        ]
      },
      "UserOut": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "education": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Education"
          },
          "direction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direction"
          },
          "goal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goal"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "max_user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max User Id"
          },
          "target_role_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Role Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at"
        ],
        "title": "UserOut",
        "description": "Профиль пользователя (ответ API).",
        "examples": [
          {
            "created_at": "2026-09-29T10:00:00",
            "direction": "backend",
            "education": "МГТУ, 3 курс",
            "goal": "Попасть на стажировку",
            "id": 1,
            "max_user_id": 12345,
            "name": "Иван Иванов",
            "target_role_id": 1
          }
        ]
      },
      "UserProgressOut": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "total_xp": {
            "type": "integer",
            "title": "Total Xp"
          },
          "completed_missions": {
            "type": "integer",
            "title": "Completed Missions"
          },
          "total_missions": {
            "type": "integer",
            "title": "Total Missions"
          },
          "skills_count": {
            "type": "integer",
            "title": "Skills Count"
          },
          "average_level": {
            "type": "number",
            "title": "Average Level"
          },
          "skill_map": {
            "items": {
              "$ref": "#/components/schemas/SkillMapItem"
            },
            "type": "array",
            "title": "Skill Map"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "total_xp",
          "completed_missions",
          "total_missions",
          "skills_count",
          "average_level",
          "skill_map"
        ],
        "title": "UserProgressOut",
        "description": "Прогресс пользователя: XP, миссии, средний уровень, Skill Map.",
        "examples": [
          {
            "average_level": 1.67,
            "completed_missions": 3,
            "skill_map": [
              {
                "category": "Программирование",
                "experience": 45,
                "level": 2,
                "name": "Python",
                "next_level_xp": 30,
                "progress": 0.75,
                "skill_id": 1
              }
            ],
            "skills_count": 3,
            "total_missions": 43,
            "total_xp": 85,
            "user_id": 1
          }
        ]
      },
      "UserUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "education": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Education"
          },
          "direction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direction"
          },
          "goal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goal"
          }
        },
        "type": "object",
        "title": "UserUpdate",
        "description": "Обновление профиля пользователя.",
        "examples": [
          {
            "direction": "frontend",
            "name": "Иван Петров"
          }
        ]
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}